Display corruption: Removed a display write
It seems like this display write contributed to the display corruption
as the problem occours much less often after removing it.

Why? I don't know... I've tried to provoke the display corruption by
writing a lot of messages in just a second, but it doesn't corrupt the
display. It therefor seems unlikely, but my only theory is still that
quick writes in succession can cause the display corruption. Although I
think it must be in combination with something else...