Feat: Make configuration easier with yaml
Instead of using an env file. Env variables are both confusing, as each
new shell has to load the variables, as well as their flat structure is
less organized to set up.

The env variables can also conflict with the user's already configured
variables and therefor had to be prefixed.