Fix: Switched to a polling rotary handler
Can't seem to get these interrupts to work optimally. They behave almost
perfectly when run on their own, but when they are mixed in with the
rest of the code, they get totally unreliable. Will have to figure this
out later.

Would seem to be some kind of performance or timing issue. Is there some
way to increase the interrupts priority to make it fire faster, maybe?